Allow menus to change content
PPS may change meals throughout the month.
When we save PDFs, hash the contents in case we get
a different pdf later on.

Include this hash in pdf/png/csv filenames,
and use only the latest hash when calculating the ics feeds.

Instead of looking to see if we already downloaded the menu
based on filename (which now includes a hash of the contents),
we store the etag on fetch, and use If-None-Match when fetching
the next time. A 304 means we can skip the download.
Since the etag can change, we also hash the body and noop
if the pdf bytes did not change.

The etag behavior is handled by cloudflare
so we can trust it to work correctly.

Also, fix a bug with prompts (we were not sending the right string),
and assert after the LLM runs that it actually returned a CSV