Add inline ics file paths
These may be needed for things like Google Calendar imports.
I am still trying to figure out why I cannot add calendar link like
https://calendar.google.com/calendar/u/0/r?cid=https://ppsmenus.net/static/feeds/lunch_k5_k8_ms.ics

It adds manually but not automatically, maybe the content disposition
of attachment is the problem, or not having an explicit content type
(netlify uses text/calendar with a charset, is that an issue maybe).